Fix Test deploys by increasing the Helm timeout to 10 minutes
We have been seeing the deployments to our test environment increasingly fail with a timeout warning[2]:

```
Error: UPGRADE FAILED: timed out waiting for the condition
```

It would be nice to properly understand what was taking too long and why it was on the test environment. Our attempts to glean more clues by using Helm haven’t found anything.

We now use the `helm_timeout` option[1] when deploying to test to give it 10 minutes instead of the default of 5 minutes.

Without a full understanding we may eventually need to add this parameter to the deploy env tasks for our other environments.

[1] https://github.com/ministryofjustice/hmpps-circleci-orb/blob/0413e8a03e25a56ea9cafee5c59c856ab40215b8/src/jobs/deploy_env.yml#L61C3-L61C15
[2] https://app.circleci.com/pipelines/github/ministryofjustice/hmpps-temporary-accommodation-ui/3723/workflows/c693d763-3636-424d-9333-12df40897db4/jobs/8691