Introduce ApplicationSummaryData type
We need to use data from the referral on pages other than the full
referral page. As the full referral page renders it's content from a
JSON object, we decided it'd be unwieldly to rely on deeply nested keys
which tightly coupled to the question copy we present in the initial
referral form.

To get around this, we are sending a `summaryData` object on submission
of the referral which will contain the summary data needed to be
presented on other pages. We still don't get true type safety, but the
object is one level deep, meaning extraction of data is more
straight-forward.

This commit adds the ApplicationSummaryData type, which we will add to
in the following commits.