Introduce anonymiseFormContent utility
In a previous commit[1] we made a change to how we save some of the form
responses by rewriting the questions that contained a person's name to
not include it. This was to ensure we weren't saving person names in the
translated document.

However, this means we have the question copy twice; one we render to
the user and one we save as part of the translated document. To avoid
confusion in the future, we will use a function to transform the
original question instead of duplicating it.

This commit introduces this function and will be used in the upcoming
commit.

We first check if the person is not an LAO (limited access offender), as
in that case, the name would already be redacted[2].

[1]
https://github.com/ministryofjustice/hmpps-temporary-accommodation-ui/pull/546

[2]
https://github.com/ministryofjustice/hmpps-temporary-accommodation-ui/pull/574/commits/539f3249e76950f0cc20bc7045ff40ed3f70a0ab