Use anonymiseFormContent in forms
In a previous PR[1], we added anonymised versions of questions for
the benefit of the translated document. This created a duplicate of
questions/titles which could lead to bugs in future if someone misses
it.

This commit makes use of the newly introduced anonymiseFormContent util,
so we can reuse the original question/title.

We also standardise language by using "the person" instead of "this
person" as we use "the person" whereever we check for LAOs[2] and this
is also in line with CAS1.

[1]
https://github.com/ministryofjustice/hmpps-temporary-accommodation-ui/pull/546

[2]
https://github.com/ministryofjustice/hmpps-temporary-accommodation-ui/commit/539f3249e76950f0cc20bc7045ff40ed3f70a0ab