Remove person name from form page titles
Presently we reuse the page title for the html document title on form
pages. As some of the page titles contain a person's name, this is a
security issue.

This commit decouples the form page title from what's applied as the
html document title by introducing a new required `htmlDocumentTitle`
property.

We also change the titles which contain a person's name and instead put
"the person" it's place.