Ensure premise is active in e2e edit premises test
In an earlier commit[1] we changed the success message when updating a
property so that if the property was archived, it would specify that in
the message.

This change ensures that we always edit an active premises in the e2e
test so we can assert the correct success message.

[1]
https://github.com/ministryofjustice/hmpps-temporary-accommodation-ui/commit/fd430c40eb09ee199c93cf6dd5873c84fb7489d4