Handle API returning 404 for ACCT alerts
In the event the API returns a 404 we would throw an unhandled exception
which would stop the user completing the referral.

This change ensures that we handle this scenario and throw for anything
else.

The UI already handles the scenario of the acctAlerts array being empty,
we so we pass an empty array in this case.