Rename Temporary Accommodation to Transitional Accommodation in UI
It's been decided that the Temporary Accommodation service will now be
called the Transitional Accommodation service. Therefore we need to
update any places that reference the old name.

In this change, we update all the places in the user interface as this is
essential for rolling out the service to the rest of the regions.

We will address the non-user-facing changes in separate commits as these
are not as urgent.