Render accommodation sharing in tracking banner
In a previous change[1], we included presentation data in the initial
submission of a referral, for easier extraction in the assess & manage
journeys.

This commit adds the sharing suitability information to the tracking
banner if it's available.

As we're accessing this data from a free object with no type safety, we
explicitly check for `true` or `false`. If this data is not present, the
conditional will return falsey and not render the section.

[1]
https://github.com/ministryofjustice/hmpps-temporary-accommodation-ui/pull/620