Introduce summaryData test fixture
In a previous change[1], we included presentation data in the initial
submission of a referral, for easier extraction in the assess & manage
journeys.

We want to be able to check for the presence of data we extract from the
`summaryData` object in our integeration & e2e tests. This commit adds a
test fixture for this data.

[1]
https://github.com/ministryofjustice/hmpps-temporary-accommodation-ui/pull/620