Send duty of care submitted data with application
We want to add data to the MI report. This commit adds data for whether
the duty of care has been submitted at the time of the application.

We follow the pattern set by arrivalDate[1] and throw an error if this
data is missing. This will guard against changing the associated page
name and not updating the reporting logic as the integration tests will
fail.

[1]
https://github.com/ministryofjustice/hmpps-temporary-accommodation-ui/blob/main/server/form-pages/utils/index.ts#L208