Send application eligibility data with application
We want to add data to the MI report. This commit adds application
eligibility data as a first class property when submitting an
application.

If the reason matches one of the 4 reasons from the application it'll
return true, which should be always. However, to guard against future
updates, we return false if the reason is not found.