Show offence ID on check your answers and full referral pages
HPTs currently have no way of knowing which index offence a referral
belongs to and are having to ask the CPP, therefore, we want to show the
offence ID on the referral page under the 'sentence information' section.

To achieve this we are adding the offence ID to the response of the
first page of the 'sentence information' task. As there is no way to
change the offence ID without starting a new referral, we hide the
"change" action on the check your answers page.

This isn't an ideal solution as it doesn't follow the conventions of the
other form pages, but here is some rationale:

* Offence ID is fetched by the server without the need for user input.
  We only ask the user to choose an offence ID if more than one offence
  is retrieved from the API. This means making a TaskListPage for
  offence ID non-trivial and would require a different approach to what we
  have now. Given there are limited dev days left on CAS3, taking on this
  work isn't feasible.

* I thought about adding a new type of TaskListPage which wouldn't be
  visible to the user, but changing the existing domain model for this
  one instance felt a bit premature, and would still be a good chunk of
  work.

* The solution in this commit is a hack, but is fairly non-invasive and
  easy to change in the future.