Show submitted referrals
In a previous change[1], we hid submitted referrals due to it causing
confusion for users. We were linking the user back to the check your
answers page which gave the impression that the application wasn't
submitted.

We are now going to link these referrals to a new page which shows the
full submitted referral.

This commit reinstates the submitted referrals tab and renders a list of
submitted referrals.

[1]
https://github.com/ministryofjustice/hmpps-temporary-accommodation-ui/pull/516