Change arrival date column to date submitted
In a previous commit[1], the shape of application data changed and as
result, we stopped showing the arrival date. This is due to the logic
relying on the 'basic-information' key which was removed.

Fixing this requires API changes to return the arrivalDate on submitted
applications, and due to the limited dev days left on CAS3, this isn't
possible right now.

As we have added the tab for submitted referrals back, we are changing
the column to instead show the date submitted and hide the column for in
progress referrals.

In our e2e test for checking the submitted application, we add a flag
for checking the submittedAt date. This is because it is set by the API
so it is difficult to assert in our e2e test. When running our
integration test however, we have control over which applications appear
so we still assert in that context.

[1]
https://github.com/ministryofjustice/hmpps-temporary-accommodation-ui/commit/ec0a2b80ce98afc24319e480f7610a048160916f