Add print test helpers to shared Page component
We want to be able to assert that the print button is present and
working on any page it's included on. This change adds these methods to
the parent Page component which other test pages inherit from.

We also want to ensure we are only testing that the `window.print()`
method was called in our integration tests.