Rename domain config to support a second
We’ll soon be adding a new domain name to the service. Before we do we rename the previous variable such that we can end up with `firstDomain` and `secondDomain`.

We’ve already gone live with the value of `firstDomain` so we’ll have to offer backwards support for it for the foreseeble.