Enable the new domain in test
Before this change transitional-accommodation-test.hmpps.service.justice.gov.uk will not resolve.

We have new test url provisioned as an allowed hostname on the certificate[1].

This change should route requests to that domain through to this service.

When we’re happy it works in test and signing in and out will redirect the user to the new domain, we can turn it on in all envs.

[1] https://github.com/ministryofjustice/cloud-platform-environments/pull/17014