Introduce userHasReporterRole util
We want to add a new 'reporter' role which will enable reporters to
download reports for all regions. This commit introduces a way to check
that the current user has this role assigned.

As the work on the API hasn't been started yet we do not have updated
types, so we're using type assertion to satisfy the compiler.