Add Oauth strategy to allow sign in attempts from new domain
We use the passport library to create a new strategy for the second domain. When passport is authenticating requests via passport.authenticate() it should cycle through the availabe strategies to check for a match to the auth provider.

Testing this locally was a fiddly. We have some branches set up on this repo[1] and our approved-premises-tools repo[2] to simulate the error before this change and how it works after being applied.

With those branches checked out (rather than this one) you can run `docker-compose down -v && tilt up -- --local-api --local-ui` within ap-tools to test.

In this test we aim to replicate the current state of the test environment where we observe an “Authorisation error”. We start two frontend instances on localhost:3000 and 127.0.0.2:3002. This simulates our service running on two different domains: https://temporary-accommodation-test.hmpps.service.justice.gov.uk and https://transitional-accommodation-test.hmpps.service.justice.gov.uk.

We check first we can sign into localhost:3000 as normal. This should test that the original mechanism is still functional. We then sign out.

Finally we visit 127.0.0.2:3002 which is the ‘new’ domain and see if we can sign in without issue. it also shows that we are redirected to 127.0.0.2:3002 rather than the original localhost:3000, which leaves us open to adding a redirect for anyone using the old url.

[1] https://github.com/ministryofjustice/hmpps-temporary-accommodation-ui/tree/test-multiple-domain-sign-in
[2] https://github.com/ministryofjustice/hmpps-approved-premises-tools/tree/test-multiple-cas3-sign-in