Update copy for when RoSH info can't be retrieved
When RoSH information can't be retrieved, we show guidance to the CPPs
to manually copy the information from Oasys and include it in the
placement consideration risk sections.

However, the guidance for HPTs is for them to manually check outside the
service so they wouldn't know to check that location for the manually
included information.

This change updates the copy in the RoSH section to prompt HPTs to look
in the placement considerations section for this information.