Introduce GetReferenceDataService
We currently make use of the referenceDataClient in multiple services.
We now need to fetch local authority reference data on the Duty To
Refer form page.

It's now a good time to introduce a reference data service to be used in
this instance and can be used across the other services in further
refactors.