Allow Reporting endpoints more time to respond
We’re finding these endpoints can take more than 30 seconds to respond. This is understandable as the data they’re loading on demand is big.

A longer term solution will be to move this reports from being on demand and to being generated in advance. This will take more time that we have left to do at this point so we try to find a compromise.

The timeout of RestClient doesn’t seem configurable from the reports controller. This might have be a cleaner option rather than repeating the API config.