Update BedspacesController to conditionally update room name
We want to allow users to be able to change a bedspace reference (room
name). As of right now, the API will fail with a bad request if a
bedspace is updated with the same name as before as it has to be unique.
To get around this, we check if the name matches and send a null value
when updating the bedspace.