Temporarily add authentication logging to help debug auth issues
On request of the HAAR team we temporarily add in this debugging so that we can assist them debug Auth issues us and other services have been experiencing. This approach is lifted from another team who have already added it[1] where we redefine the implementation of the callback `successReturnToOrRedirect` with our custom handling of the same.

We make one change from the implementation we’re copying from: we guard against the absence of info.message and req.session.returnTo as we can’t be sure they’ll always be there.

We’re seeing lots of autherrors and have been advising users to clear their cookies.

[1] https://github.com/ministryofjustice/hmpps-digital-prison-services/pull/84