Remove testing mechanism to create bulk bookings
We changed approach so that the API will seed these bookings upfront much more quickly. We can’t foresee that the e2e tests will need to do this since we’d run into the same efficiency problems.

The comment below points to the need to keep it for future use. This is well intended but we should remove this complexity from the app until it’s actually needed. In which case this code can be easily re-proposed.

```
// The bookings are now populated by the api in dev, but leaving the bulk() helper for
// future use (thinking performance testing)
```