Split existing tabbed index into separate pages
This commit only creates and tests the new URLs, while the call to the
API is still requesting all assessments at once: this allows the
splitting of pages to be done independently from API requests changes.

A redirect from the old index path (/review-and-asses) has been put in
place to ensure tests pass with minimal change, but also to prevent
breaking a user's workflow should they have bookmarked the original url.