Ignore old release types in Review your answers
The change of release types prevented displaying the 'Review your
answers' page if an old application contained one of the old release
types, 'licence'. This one is now filtered out of the possible answers.
If the resulting application has no release type, it will now show as
incomplete and a session error would be thrown if accessing the 'Review
your answers' page, which is correct for all applictaion errors.