Fix check for past and future dates
Original implementation relied on a third-party library, which compared
_datetimes_ rather than _dates_, meaning that, in most cases, today's
date was considered in the past.

For this service, the only useful comparison is for dates only.

The original implementation was also relying on tests that simply mocked
the boolean value return from the third-party library, hiding the issue.

This implementation now relies on stubbing system time, as is
recommended by Jest.