Render full referral with updated dates
Leaving the overriding of submitted data with updated data to the
template level allows us to choose whether the document should be shown
as submitted or not. This comes at the price of readability, and should
most likely be revisited with a more robust solution for _any_ kind of
data update.