fix report button styling
I noticed that the report button styling was not consistent with the rest of the application. I discovered that `govuk-input` class was incorrectly added to them. This meant that the buttons would have a thicker border around them like text inputs and it also changed how the button looked like when the mouse hovered over it. This change would more than likely have broken the contrast accessibility for these buttons.