Fix pages with single date inputs
The current implementation causes a lot of accesibility problems for screen readers as the screenreader does not have context as to what is being asked from the user. All it will read out is "Day", "Month", or "Year"  depending on which input has focus. We should be using IsPageHeading option so that a fieldset is placed around the date field but it also injects an h1 tag inside the fieldsets legend. A screen will then announce the legend to the user so the will have the qhestion and hint

THere was an accessibility issue reported as part of an assessment done late 2023 which read:

>Date input fields are not grouped with questions, so screen reader users may find it challenging to associate the date input field with the relevant question (test issue 1.3). This fails WCAG 2.2 success criterion 1.3.1 (A). We plan to fix this by 30/03 2024.

This commit should mean this issue is now fixed across the service.