Rename search crn param to query
We are changing how search works and in do so the search param will no longer only search for crn but name as well.
This commit is to rename the existing param so its clear that its not only for crn. The factory was also updated
from a CRN to a random sentenance to show that the query param could potentially be used for string and not just CRN