Add SearchByCrnOrNameResult component
This commit moves majority of the search form results into its own component. While refactoring the following inconsistency were discovered between referral and bookings.

referral displays pagination above and below the results where as booking only displays it at the bottom. We will need to consult with designers to confirm the layout.

Bookings results table includes captionClasses attribute which is not actually being used and could be removed in the future commits.  Its also missing `firstCellIsHeader` seting to make the persons name part of the row heading. Screen readers will call out the column heading, row heading followed by the data the user is currently selected. This improves accessibilities and helps users navigate the table.

As there are these inconsistencies between the layout we will update the component later on so the component will be responsible for building the all the results and not the view.