Update alternative PDU form page
This provides a dropdown to select the alternative PDU with all values
available from the DB instead of a free text box. The details question
is now on a separate page, which is only shown if the user has selected
'Yes' to the main answer. The 'Review your answers' and referral summary
screens are updated to show the relevant answers only.