Always show persons Sex and conditionally display Gender Identity
In nDelius, gender identity is not a required field but if the user has been provided with an answer then they have to record it. Previously we were displaying either the persons sex or their gender identity which was causing issues for our users especially when trying to find a bedspace and trying to identify if the person is male/female.  So now we will also display the persons biological sex and display gender identity even it has been answered. This includes instances where gender identity and sex match or if they have answers "Prefer not to say".