Add notification banner with link to external guidance to "Alternative PDU" page
Background: The central HPT team and HPT Leads want to flag that accommodation is not guaranteed when PPs request accommodation in a PDU outside of their probation region. Due to policy change and increase in number of referrals the central HPT team wants point PPs to relevant guidance on Equip if PP want to make an out of region referral.

Acceptance Criteria

A warning message displayed on ‘Is accommodation required in alternate PDU section’ on the referral form

The warning message should inform PP to check guidance and speak to relevant HPT team if accommodation is required in an alternate PDU outside of the PP’s region.