Update getActingUserPrimaryNavigationList to accept user and currentpage
Based on the users role, if they are an assessor they will get a list of links that they can visit. Otherwise, and empty array is returned.

Depending on the currentpage the user is viewing, if it contains part of the url that the nav link is then that nav link will display as being active.