Ensure practitioner PDU test uses distinct ids
When buildung a list, the PDU factory may in some cases return 2 of the
same PDU from the available fixtures. This means the test for
getRegionsPdus would sometimes produce 2 similar entries which would
both be marked as selected, which would fail the test. This ensures
every PDU in the test has a different id, preventing this issue.