Centralise redirects to be routes
Previously, we had a couple controller methods created which were purely for redirecting to another page. This is a code smell and could be done much easier by using a generic redirect controller.

We now have this controller in place so this commit is to update the previous pattern to now all be redirected straight from routes and not in the individual controllers.