Allow users to search bedspaces for up to 5 PDUs
Previously users could only search for bedspaces for a single
PDU. This was very limiting and so we extended this up to 5 PDUs.

The reason for the limit of 5 was because of a previous tech spike
into the performance of the API querie and found that anything more
would cause poor performance. See https://dsdmoj.atlassian.net/browse/CAS-315