Fix bug where assessments return all ready for placement
This was caused by a recent change where we started allowing users
to search for an assessment by crn or name. The backend api params
changed from crn to crnOrName. As a result this change it introduced
this bug of returning all assessments that were ready for placement
instead of all assessments for a given crn that were ready for placement.