Use the PDU ID that is on the premises show page
the apply-and-place feature tests were failing when the feature got
to "Search for available bedspace" page because it could not find a
checkbox with a pdu id from an existing premises. After debugging the
tests it was discovered that we had set the premises id as the PDU id
when creating the PDU from a factory.

apply-andplace features use the ui to create the various
assessments/premises etc and scrape the screen to get enough
information to use with factories to store the data in memory
for the tests to use dynamically. The problem with this approach
is that the pdu id was not surfaced to the ui at all so could not
use it when building the pdu.

I decided to surface the pdu id on the premises page where most of
the information about the premises is surfaced and place it in
a html data attribute called `data-cy-pdu-id`. The tests can now
read that id from the page and store it for later use on the
search available bedspaces page