Restrict the creation of application using obsolete release types
We need to remove ECSL as a release type option from 10 Sept as ECSL will no longer exist. We are aware that there may still be some applications being processed so instead of simply removing the value from the field, we decided to instead support obsolete release types if the application previous had ECSL we would still show those options in the radio list unless the user updated it to so other release. This approach means that we can slowly depreciate the release types and over time easiely drop the release type altogether.