Integrate TimelineService into PP Application Controller
We need to display the assessment timeline to PP but not include HPT notes for the time being. The restriction of the HPT notes not being shown to PP is done in the API and so the UI has no control of it. This commit modifies the applicationsController.ts to call the new TimelineService and to pass the results back to the view to display the timeline to PP.