Replace custom datepicker with official moj date picker
We were one of the original services testing the datepicker for MOJ design system team. This component has since been included in MoJ Frontend so we no longer need to maintain this custom code.

There are some minor visual differences, eg. calendar appears below the input instead of to the right of it. I dont think we should be concerned about this as this is now the official component and if any browser/accessibility issues crop up then it should be fixed upstream and find its way to our service.