Reorganise api definitions to introduce cas3v2
This renames a lot of the transitory variables to be prefixed with
`cas3` or `cas3v2` where necessary, mostly for readability. It also
moves any path that is already querying the `/cas3` namespace on the
api to the `cas3` property, again, for clarity.

Finally, this introduces the feature-switchable `/cas3/v2` prefix for
premises and bedspaces endpoints that are already available on this
namespace.