Fix issue with samesite cookies inside Cypress UI
A recent change to the Auth service specifies `samesite=Strict` for auth
cookies. This breaks flows in Cypress UI, as those are run inside a
frame for the given URL, which means they never pass the samesite
validation. This change intercepts requests containing those auth
cookies to remove the restriction.