Use unique references for bedspace search results
In some rare casess, multiple search results as generated for the
integration tests may have had references starting with the same
numbers, as the references themselves were chosen as random numbers
between 1 and 9999.

For instance, one result could have had reference '81', with
another having reference '8190'. In this case, the integration test
searching for teh result with reference '81' may have returned the
result container for the one with reference '8190', breaking the
assertions further down in the test.

To fix this, we use faker's alphanumeric string generator for results
reference, to minimise the chance of collisions.